requirements.txt
pandas>=2.0
numpy>=1.24
biopython>=1.83
requests>=2.31
mmh3>=4.1.0
